Add files via upload
The Educational Institutions or any other organisations need to monitor the movements of individuals within their boundaries. This task cannot be handled manually on a consistent basis. In order to assist the human force, we can use the technical systems.We have the attendance monitoring systems based on face recognition, biometric, blue-tooth, Rf-id and many more.In this project, we have considered Rf-id based attendance monitoring system which can help the teaching staff. This project uses Rf-id reader which reads when the Rf-id tag is brought nearer.Then the reader processes and sends the data to the micro-controller which stores the values and displays the desired output on the lcd. Once the authenticator presses the tactile switch, the micro-controller sends the data to the server(ThingSpeak) through the wifi device(esp8266).The attendance of the individuals will be displayed.This project’s main motive is to aid teachers’ in managing the attendance of the students.